change design for visitor part 2
