+++ aderyn_core/src/visitor/mod.rs
@@ -1,1 +1,3 @@
 pub mod ast_visitor;
+pub mod macros;
+pub mod workspace_visitor;
